Jest preset fixes/improvements (#501)
### 🔦 Summary

This PR introduces a few changes:

* It replaces how you start a local dev server: instead of supplying a path to a file, you supply which command to spawn (e.g. `node index.js`). 
* It also uses default config values instead of throwing an error if a `jest-yoshi.config.js` was not found.
* Starting a local server/launching puppeteer is only done if any puppeteer tests are actually going to run.

+++ packages/jest-environment-yoshi-puppeteer/globalTeardown.js
@@ -1,12 +1,15 @@
 const fs = require('fs-extra');
 const { WS_ENDPOINT_PATH } = require('./constants');
+const { shouldRunE2Es } = require('./utils');
 
 module.exports = async () => {
-  await fs.remove(WS_ENDPOINT_PATH);
+  if (await shouldRunE2Es()) {
+    await fs.remove(WS_ENDPOINT_PATH);
 
-  await global.BROWSER.close();
+    await global.BROWSER.close();
 
-  if (global.SERVER) {
-    global.SERVER.kill();
+    if (global.SERVER) {
+      global.SERVER.kill();
+    }
   }
 };
